+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/TigerProxyApplication.java
@@ -4,10 +4,16 @@
 
 package de.gematik.test.tiger.proxy;
 
+import de.gematik.rbellogger.captures.RbelFileReaderCapturer;
 import de.gematik.test.tiger.common.data.config.tigerProxy.TigerProxyConfiguration;
 import de.gematik.test.tiger.proxy.configuration.ApplicationConfiguration;
+import java.io.File;
+import java.util.Base64;
+import java.util.stream.Stream;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import org.apache.commons.io.FileUtils;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
@@ -26,6 +32,7 @@ public class TigerProxyApplication {
         SpringApplication.run(TigerProxyApplication.class, args);
     }
 
+    @SneakyThrows
     @Bean
     public TigerProxy tigerProxy() {
         if (applicationConfiguration != null) {